+++ product-manage/src/main/java/com/example/demo/entity/dto/response/ProductPurchaseResponseDto.java
@@ -1,5 +1,7 @@
 package com.example.demo.entity.dto.response;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
@@ -19,7 +21,7 @@ public class ProductPurchaseResponseDto {
 
 	private String productColor;
 
-	private Integer productPurchaseUnitPrice;
+	private BigDecimal productPurchaseUnitPrice;
 
 	private Integer productStockQuantity;
